Using GLshort for vertex coordinates instead of GLfloat, bugfix in BltTriangles

+++ src/lib/D3DInterface.h
@@ -23,8 +23,8 @@ class TriVertex;
 #ifndef WIN32
 //Aligned vertex structure
 typedef struct {
-    GLfloat sx;
-    GLfloat sy;
+    GLshort sx;
+    GLshort sy;
     SexyRGBA color;
     GLfloat tu;
     GLfloat tv;
